+++ snake.h
@@ -32,15 +32,15 @@ void generateObjects();
 void outputSnakePlayfield();
 void endGame();
 uint8_t addSnakeVertex(uint8_t x, uint8_t y);
-uint8_t getSnakeVertexX(int index);
-uint8_t getSnakeVertexY(int index);
+uint8_t getSnakeVertexX(int16_t index);
+uint8_t getSnakeVertexY(int16_t index);
 void resetSnake();
 void renderSnake();
 t_collisionType advanceSnake();
 void feedSnake(uint8_t amount);
-unsigned int getSnakeDistance(int x0, int y0, int x1, int y1);
-t_direction getSnakeDirection(int x0, int y0, int x1, int y1);
-uint8_t checkForWall(int x, int y);
-uint8_t checkForSnake(int x, int y);
+uint16_t getSnakeDistance(int16_t x0, int16_t y0, int16_t x1, int16_t y1);
+t_direction getSnakeDirection(int16_t x0, int16_t y0, int16_t x1, int16_t y1);
+uint8_t checkForWall(int16_t x, int16_t y);
+uint8_t checkForSnake(int16_t x, int16_t y);
 
 #endif /* SNAKE_H_ */
